refactor types to int, fix sale logic and seed data

+++ Domain/CashDesk/SaleItem.cs
@@ -4,10 +4,10 @@ public class  SaleItem
 {
     public string Barcode { get; private set; }
     public string Name { get; private set; }
-    public long Price { get; private set; }
-    public long Quantity { get; set; }
+    public int Price { get; private set; }
+    public int Quantity { get; set; }
 
-    public SaleItem(string barcode, string name, long price)
+    public SaleItem(string barcode, string name, int price)
     {
         Barcode = barcode;
         Name = name;
